extract hooks into its own module
