+++ docs/api/nn/Reshape.md
@@ -1,0 +1,9 @@
+# elegy.nn.Reshape
+
+::: elegy.nn.Reshape
+    selection:
+        inherited_members: true
+        members:
+            - call
+            - __init__
+            - __call__
