+++ elegy/__init__.py
@@ -2,6 +2,15 @@ __version__ = "0.1.5"
 
 
 from . import callbacks, initializers, losses, metrics, model, nn, regularizers
+from .hooks import (
+    add_loss,
+    add_metric,
+    add_summary,
+    get_parameter,
+    get_state,
+    next_rng_key,
+    set_state,
+)
 from .losses import Loss
 from .metrics import Metric
 from .model import Model
@@ -11,13 +20,6 @@ from .module import (
     Context,
     InitCallable,
     Module,
-    add_loss,
-    add_metric,
-    get_parameter,
-    get_state,
-    add_summary,
     context,
-    next_rng_key,
     to_module,
-    set_state,
 )
