+++ elegy/initializers.py
@@ -15,7 +15,7 @@
 # ==============================================================================
 """Haiku initializers."""
 
-from elegy import module
+from elegy import module, hooks
 from elegy.types import DType, Initializer, Shape
 import jax
 import jax.numpy as jnp
@@ -72,7 +72,7 @@ class RandomNormal(Initializer):
     def __call__(self, shape: Shape, dtype: DType) -> np.ndarray:
         m = jax.lax.convert_element_type(self.mean, dtype)
         s = jax.lax.convert_element_type(self.stddev, dtype)
-        return m + s * jax.random.normal(module.next_rng_key(), shape, dtype)
+        return m + s * jax.random.normal(hooks.next_rng_key(), shape, dtype)
 
 
 class TruncatedNormal(Initializer):
@@ -94,7 +94,7 @@ class TruncatedNormal(Initializer):
         m = jax.lax.convert_element_type(self.mean, dtype)
         s = jax.lax.convert_element_type(self.stddev, dtype)
         unscaled = jax.random.truncated_normal(
-            module.next_rng_key(), -2.0, 2.0, shape, dtype
+            hooks.next_rng_key(), -2.0, 2.0, shape, dtype
         )
         return s * unscaled + m
 
@@ -115,7 +115,7 @@ class RandomUniform(Initializer):
 
     def __call__(self, shape: Shape, dtype: DType) -> np.ndarray:
         return jax.random.uniform(
-            module.next_rng_key(), shape, dtype, self.minval, self.maxval
+            hooks.next_rng_key(), shape, dtype, self.minval, self.maxval
         )
 
 
@@ -254,7 +254,7 @@ class Orthogonal(Initializer):
         n_rows = shape[self.axis]
         n_cols = np.prod(shape) // n_rows
         matrix_shape = (n_rows, n_cols) if n_rows > n_cols else (n_cols, n_rows)
-        norm_dst = jax.random.normal(module.next_rng_key(), matrix_shape, dtype)
+        norm_dst = jax.random.normal(hooks.next_rng_key(), matrix_shape, dtype)
         q_mat, r_mat = jnp.linalg.qr(norm_dst)
         # Enforce Q is uniformly distributed
         q_mat *= jnp.sign(jnp.diag(r_mat))
