+++ elegy/nn/__init__.py
@@ -1,7 +1,7 @@
 from .batch_normalization import BatchNormalization
 from .conv import Conv1D, Conv2D, Conv3D, ConvND
 from .dropout import Dropout
-from .flatten import Flatten
+from .flatten import Flatten, Reshape
 from .linear import Linear
 from .sequential import Sequential, sequential
 from .layer_normalization import LayerNormalization, InstanceNormalization
