+++ elegy/nn/batch_normalization.py
@@ -11,7 +11,7 @@ import jax.numpy as jnp
 import numpy as np
 from haiku._src import utils as haiku_utils
 
-from elegy import initializers, module
+from elegy import initializers, module, hooks
 from elegy import module
 from elegy.nn.moving_averages import ExponentialMovingAverage
 
@@ -120,13 +120,7 @@ class BatchNormalization(module.Module):
         The array, normalized across all but the last dimension.
         """
         if training is None:
-            if module.LOCAL.contexts:
-                context: module.Context = module.LOCAL.contexts[-1]
-                training = context.training
-            else:
-                raise ValueError(
-                    "Cannot infer `training` outside of context. You must pass `training` explicitly or use `Module.apply`."
-                )
+            training = hooks.is_training()
 
         if self.create_scale and scale is not None:
             raise ValueError("Cannot pass `scale` at call time if `create_scale=True`.")
@@ -170,12 +164,12 @@ class BatchNormalization(module.Module):
         w_dtype = inputs.dtype
 
         if self.create_scale:
-            scale = module.get_parameter("scale", w_shape, w_dtype, self.scale_init)
+            scale = hooks.get_parameter("scale", w_shape, w_dtype, self.scale_init)
         elif scale is None:
             scale = np.ones([], dtype=w_dtype)
 
         if self.create_offset:
-            offset = module.get_parameter("offset", w_shape, w_dtype, self.offset_init)
+            offset = hooks.get_parameter("offset", w_shape, w_dtype, self.offset_init)
         elif offset is None:
             offset = np.zeros([], dtype=w_dtype)
