+++ elegy/nn/conv.py
@@ -23,7 +23,7 @@ import numpy as np
 from haiku._src import utils as hk_utils
 from jax import lax
 
-from elegy import initializers, module, types
+from elegy import initializers, module, types, hooks
 
 
 def to_dimension_numbers(
@@ -170,7 +170,7 @@ class ConvND(module.Module):
             fan_in_shape = np.prod(w_shape[:-1])
             stddev = 1.0 / np.sqrt(fan_in_shape)
             w_init = initializers.TruncatedNormal(stddev=stddev)
-        w = module.get_parameter("w", w_shape, inputs.dtype, initializer=w_init)
+        w = hooks.get_parameter("w", w_shape, inputs.dtype, initializer=w_init)
 
         if self.mask is not None:
             w *= self.mask
@@ -190,7 +190,7 @@ class ConvND(module.Module):
                 bias_shape = (self.output_channels,)
             else:
                 bias_shape = (self.output_channels,) + (1,) * self.num_spatial_dims
-            b = module.get_parameter(
+            b = hooks.get_parameter(
                 "b", bias_shape, inputs.dtype, initializer=self.b_init
             )
             b = jnp.broadcast_to(b, out.shape)
@@ -481,7 +481,7 @@ class ConvNDTranspose(module.Module):
             fan_in_shape = self.kernel_shape + (input_channels,)
             stddev = 1.0 / np.sqrt(np.prod(fan_in_shape))
             w_init = initializers.TruncatedNormal(stddev=stddev)
-        w = module.get_parameter("w", w_shape, inputs.dtype, initializer=w_init)
+        w = hooks.get_parameter("w", w_shape, inputs.dtype, initializer=w_init)
 
         if self.mask is not None:
             w = w * self.mask
@@ -499,7 +499,7 @@ class ConvNDTranspose(module.Module):
                 bias_shape = (self.output_channels,)
             else:
                 bias_shape = (self.output_channels,) + (1,) * self.num_spatial_dims
-            b = module.get_parameter("b", bias_shape, initializer=self.b_init)
+            b = hooks.get_parameter("b", bias_shape, initializer=self.b_init)
             b = jnp.broadcast_to(b, out.shape)
             out = out + b
