+++ elegy/nn/dropout.py
@@ -8,7 +8,7 @@ import haiku as hk
 import jax.numpy as jnp
 import numpy as np
 
-from elegy import module
+from elegy import module, hooks
 from elegy.module import Module
 
 
@@ -68,15 +68,10 @@ class Dropout(Module):
             x but dropped out and scaled by `1 / (1 - rate)`.
         """
         if training is None:
-            if module.LOCAL.contexts:
-                context: module.Context = module.LOCAL.contexts[-1]
-                training = context.training
-            else:
-                raise ValueError(
-                    "Cannot infer `training` outside of context. You must pass `training` explicitly or use `Module.apply`."
-                )
+            training = hooks.is_training()
+
         return hk.dropout(
-            rng=rng if rng is not None else module.next_rng_key(),
+            rng=rng if rng is not None else hooks.next_rng_key(),
             rate=self.rate if training else 0.0,
             x=x,
         )
