+++ elegy/nn/flatten.py
@@ -36,45 +36,49 @@ class Reshape(Module):
     """
     Reshapes input Tensor, preserving the batch dimension.
 
-    For example, given an input tensor with shape ``[B, H, W, C, D]``::
+    For example, given an input tensor with shape `[B, H, W, C, D]`:
 
-        >>> B, H, W, C, D = range(1, 6)
-        >>> x = jnp.ones([B, H, W, C, D])
+    ```python
+    B, H, W, C, D = range(1, 6)
+    x = jnp.ones([B, H, W, C, D])
+    ```
 
-    The default behavior when ``output_shape`` is ``(-1, D)`` is to flatten
-    all dimensions between ``B`` and ``D``::
+    The default behavior when `output_shape` is `(-1, D)` is to flatten
+    all dimensions between `B` and `D`:
 
-        >>> mod = hk.Reshape(output_shape=(-1, D))
-        >>> assert mod(x).shape == (B, H*W*C, D)
+    ```python
+    mod = elegy.nn.Reshape(output_shape=(-1, D))
+    assert mod(x).shape == (B, H*W*C, D)
+    ```
 
     You can change the number of preserved leading dimensions via
-    ``preserve_dims``::
+    `preserve_dims`:
 
-        >>> mod = hk.Reshape(output_shape=(-1, D), preserve_dims=2)
-        >>> assert mod(x).shape == (B, H, W*C, D)
-
-        >>> mod = hk.Reshape(output_shape=(-1, D), preserve_dims=3)
-        >>> assert mod(x).shape == (B, H, W, C, D)
-
-        >>> mod = hk.Reshape(output_shape=(-1, D), preserve_dims=4)
-        >>> assert mod(x).shape == (B, H, W, C, 1, D)
+    ```python
+    mod = elegy.nn.Reshape(output_shape=(-1, D), preserve_dims=2)
+    assert mod(x).shape == (B, H, W*C, D)
+    mod = elegy.nn.Reshape(output_shape=(-1, D), preserve_dims=3)
+    assert mod(x).shape == (B, H, W, C, D)
+    mod = elegy.nn.Reshape(output_shape=(-1, D), preserve_dims=4)
+    assert mod(x).shape == (B, H, W, C, 1, D)
+    ```
     """
 
     def __init__(self, output_shape: types.Shape, preserve_dims: int = 1, **kwargs):
         """
-        Constructs a ``Reshape`` module.
+        Constructs a `Reshape` module.
 
         Args:
             output_shape: Shape to reshape the input tensor to while preserving its
-                first ``preserve_dims` dimensions. When the special value -1 appears in
-                ``output_shape`` the corresponding size is automatically inferred. Note
-                that -1 can only appear once in ``output_shape``.
-                To flatten all non-batch dimensions use :class:`Flatten`.
+                first `preserve_dims` dimensions. When the special value -1 appears in
+                `output_shape` the corresponding size is automatically inferred. Note
+                that -1 can only appear once in `output_shape`.
+                To flatten all non-batch dimensions use `Flatten`.
             preserve_dims: Number of leading dimensions that will not be reshaped.
             kwargs: Additional keyword arguments passed to Module.
 
         Raises:
-            ValueError: If ``preserve_dims`` is not positive.
+            ValueError: If `preserve_dims` is not positive.
         """
         super().__init__(**kwargs)
         if preserve_dims <= 0:
@@ -85,7 +89,14 @@ class Reshape(Module):
         self.output_shape = tuple(output_shape)
         self.preserve_dims = preserve_dims
 
-    def call(self, inputs):
+    def call(self, inputs: np.ndarray) -> np.ndarray:
+        """
+        Arguments:
+            inputs: the array to be reshaped.
+        
+        Returns:
+            A reshaped array.
+        """
         if inputs.ndim <= self.preserve_dims:
             return inputs
 
@@ -106,17 +117,19 @@ class Flatten(Reshape):
     By default, Flatten combines all dimensions except the first.
     Additional leading dimensions can be preserved by setting preserve_dims.
 
-    >>> x = jnp.ones([3, 2, 4])
-    >>> flat = hk.Flatten()
-    >>> flat(x).shape
-    (3, 8)
+    ```python
+    x = jnp.ones([3, 2, 4])
+    flat = elegy.nn.Flatten()
+    assert flat(x).shape == (3, 8)
+    ```
 
-    When the input to flatten has fewer than ``preserve_dims`` dimensions it is
+    When the input to flatten has fewer than `preserve_dims` dimensions it is
     returned unchanged:
 
-    >>> x = jnp.ones([3])
-    >>> flat(x).shape
-    (3,)
+    ```python
+    x = jnp.ones([3])
+    assert flat(x).shape == (3,)
+    ```
     """
 
     def __init__(
