+++ elegy/nn/linear.py
@@ -1,6 +1,6 @@
 from elegy.initializers import TruncatedNormal
 from elegy.types import Initializer
-from elegy import module
+from elegy import module, hooks
 import typing as tp
 import jax.numpy as jnp
 import haiku as hk
@@ -57,14 +57,14 @@ class Linear(module.Module):
             stddev = 1.0 / np.sqrt(self.input_size)
             w_init = TruncatedNormal(stddev=stddev)
 
-        w = module.get_parameter(
+        w = hooks.get_parameter(
             "w", [input_size, output_size], dtype, initializer=w_init
         )
 
         out = jnp.dot(inputs, w)
 
         if self.with_bias:
-            b = module.get_parameter(
+            b = hooks.get_parameter(
                 "b", [self.output_size], dtype, initializer=self.b_init
             )
             b = jnp.broadcast_to(b, out.shape)
